Some edits to the comments
I made some of the comments more descriptive, and got rid of some of some extra whitespace that was hanging around. Oh, and I added Ethan's name to the list of people (but I don't know his last name-- we gotta add that in) and changed my name so that it only appeared once.